+++ build_python.sh
@@ -31,7 +31,7 @@ if [ ! -d "$1" ] || [ ! -r "${LIBNAME}" ]; then
   source conan/conanbuild.sh
 
   echo ${SEP}
-  PKGS="zlib bzip2 liblzma"
+  PKGS="zlib bzip2 liblzma gettext openssl"
   echo "conan CFLAGS=${CFLAGS}"
   echo "conan LDFLAGS=${LDFLAGS}"
   echo ${SEP}
@@ -41,7 +41,6 @@ if [ ! -d "$1" ] || [ ! -r "${LIBNAME}" ]; then
   echo ${SEP}
 
   rm -rf ${PYTHON_BUILD}/Python-${PYTHON_VERSION}
-  rm -rf ${PYTHON_BUILD}/openssl
 
   mkdir -p ${PYTHON_BUILD}
   tar -C ${PYTHON_BUILD} -zxf ${PYTHON_DIST}/Python-${PYTHON_VERSION}.tgz
@@ -51,34 +50,38 @@ if [ ! -d "$1" ] || [ ! -r "${LIBNAME}" ]; then
       patch < ${PATCH}
     done
 
-    export PY_UNSUPPORTED_OPENSSL_BUILD=static
     case `uname` in
       'Linux')
-        LDFLAGS="-Wl,-z,origin -Wl,-rpath,'\$\$ORIGIN/../lib' -Wl,-Bstatic ${LDFLAGS} `pkg-config --static --libs sqlite3` -Wl,-Bdynamic"
+        LDFLAGS="-Wl,-z,origin -Wl,-rpath,'\$\$ORIGIN/../lib' ${LDFLAGS}"
         PKGS="${PKGS} sqlite3 readline"
-        export ZLIB_LIBS="-Wl,-Bstatic `pkg-config --static --libs zlib` -Wl,-Bdynamic -ldl"
+        export ZLIB_LIBS="`pkg-config --static --libs zlib` -ldl"
         export LIBFFI_LIBS="-l:libffi_pic.a -Wl,--exclude-libs,libffi_pic.a"
+        export POSIXSHMEM_LIBS="-lrt"
         ;;
       'Darwin')
-        mkdir -p ${PYTHON_BUILD}/openssl/lib
-        mkdir -p ${PYTHON_BUILD}/openssl/include
-        cp $(brew --prefix openssl@1.1)/lib/*.a ${PYTHON_BUILD}/openssl/lib
-        cp -r $(brew --prefix openssl@1.1)/include/openssl ${PYTHON_BUILD}/openssl/include
-        mkdir -p ${PYTHON_BUILD}/gettext/lib
-        cp $(brew --prefix gettext)/lib/*.a ${PYTHON_BUILD}/gettext/lib
-        export SSL="--with-openssl=${PYTHON_BUILD}/openssl"
-        LDFLAGS="-Wl,-search_paths_first -L${PYTHON_BUILD}/gettext/lib -Wl,-rpath,@loader_path/../lib"
+        LDFLAGS="-Wl,-search_paths_first -Wl,-rpath,@loader_path/../lib"
         export LIBS="-liconv -framework CoreFoundation ${LDFLAGS}"
         ;;
     esac
     export CFLAGS="`pkg-config --static --cflags ${PKGS}` ${CFLAGS}"
-    export LDFLAGS="`pkg-config --static --libs ${PKGS}` ${LDFLAGS}"
+    export LIBS="`pkg-config --static --libs ${PKGS}` ${LIBS}"
+    export LDFLAGS="${LDFLAGS}"
 
-    ./configure --prefix $1 $2 --enable-optimizations ${SSL}
+    ./configure --prefix $1 $2 --enable-optimizations
     make -j4 build_all
     make install
   )
   rm -f $1/python
   [ ! -r $1/bin/python3 ] && ln -s python${SHORT_VERSION} $1/bin/python3
   ln -s bin/python3 $1/python
+
+  # Get the curl certificates
+  curl https://curl.se/ca/cacert.pem --output $1/cacert.pem
+  mkdir -p $1/cert
+  csplit -k -f $1/cert/root- $1/cacert.pem '/END CERTIFICATE/+1' {500}
+  for CERT in $1/cert/root-*; do
+      mv ${CERT} ${CERT}.pem
+  done
+  perl `pkg-config --variable=bindir openssl`/c_rehash $1/cert
+  echo SSL_CERT_DIR=$1/cert
 fi