+++ conanfile.py
@@ -8,8 +8,14 @@ class PythonDeps(ConanFile):
     self.requires('zlib/1.3.1')
     self.requires('xz_utils/5.4.5')
     self.requires('bzip2/1.0.8')
+    self.requires('openssl/3.3.2')
+    self.requires('gettext/0.22.5')
 
     # These libraries are part of the OS on macOS, but are optional on Linux
     if self.settings.os == 'Linux':
       self.requires('sqlite3/3.46.1')
       self.requires('readline/8.2')
+
+  def configure(self):
+    self.options['sqlite3'].enable_fts3 = True
+    self.options['sqlite3'].enable_fts4 = True