+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+#
+
+- Get OpenSSL from conan and include root CA certificates from the CURL project
+
 # 2024-07-01
 
 - [#3](https://github.com/mmomtchev/static-portable-python/issues/3), on Linux, avoid dependencies on `sqlite3.so` and `libreadline.so.8`